Schueler und Notenprocessorn sollten full async sein. mit connectionstring aus appsettinngs.json lesen createschuelerasync geht ned ganz wegen 2nd psql statement

+++ Ubigrade.Library/Processors/SchuelerProcessor.cs
@@ -11,11 +11,11 @@ namespace Ubigrade.Library.Processors
 {
     public static class SchuelerProcessor
     {
-        public static void CreateSchueler(int checknp, string nname, string vname, string geschlecht, string email, int sjahr)
+        public async static Task<bool> CreateSchuelerAsync(int checknp, string nname, string vname, string geschlecht, string email, int sjahr,string sql)
         {
             try
             {
-                using (NpgsqlConnection connection = new NpgsqlConnection(SqlDataAccess.GetConnectionString()))
+                using (NpgsqlConnection connection = new NpgsqlConnection(sql))
                 {
                     NpgsqlCommand command;
                     var x = geschlecht.ToUpper();
@@ -24,16 +24,18 @@ namespace Ubigrade.Library.Processors
                     command =
                             new NpgsqlCommand
                             ($"call insertschueler('{vname}', '{nname}', '{geschlecht.ToUpper()}', '{email}', {checknp}, {sjahr});", connection);
-                    command.ExecuteNonQuery();
+                    int i = await command.ExecuteNonQueryAsync();
 
 
                     //Fehler
                     command =
                         new NpgsqlCommand
                         ($"insert into schueler2faecher (bez, skennzahl) VALUES ('M',{checknp},('E',{checknp}),('NWT1',{checknp}),('WIRE',{checknp}),('ITP2',{checknp});", connection);
-                    command.ExecuteNonQuery();
-
+                    i += await command.ExecuteNonQueryAsync();
                     connection.Close();
+                    if (i == 2)
+                        return true;
+                    return false;
                 }
             }
             catch (Exception ex)
@@ -42,11 +44,11 @@ namespace Ubigrade.Library.Processors
             }
         }
 
-        public async static Task<List<SchuelerDLModel>> LoadSchuelerAsync(/*string sql*/ )
+        public async static Task<List<SchuelerDLModel>> LoadSchuelerAsync(string sql)
         {
             List<SchuelerDLModel> ListeSchueler = new List<SchuelerDLModel>();
 
-            using (NpgsqlConnection connection = new NpgsqlConnection(SqlDataAccess.GetConnectionString()))
+            using (NpgsqlConnection connection = new NpgsqlConnection(sql))
             {
                 ListeSchueler.Clear();
                 NpgsqlCommand command;
@@ -78,9 +80,9 @@ namespace Ubigrade.Library.Processors
             return ListeSchueler;
         }
 
-        public static void SaveSchueler(int id, SchuelerDLModel changedschueler)
+        public async  static Task<bool> SaveSchuelerAsync(string sql, int id, SchuelerDLModel changedschueler)
         {
-            using (NpgsqlConnection connection = new NpgsqlConnection(SqlDataAccess.GetConnectionString()))
+            using (NpgsqlConnection connection = new NpgsqlConnection(sql))
             {
                 NpgsqlCommand command;
 
@@ -90,24 +92,23 @@ namespace Ubigrade.Library.Processors
                 new NpgsqlCommand
                 ($"update person set checkpersonnumber = {changedschueler.Checkpersonnumber}, vname = '{changedschueler.VName}', nname = '{changedschueler.NName}', " +
                 $"gender = '{changedschueler.Geschlecht.ToUpper()}', email = '{changedschueler.EmailAdresse}' where checkpersonnumber = {id}; ", connection);
-                command.ExecuteNonQuery();
-
-                connection.Close();
+                int i = await command.ExecuteNonQueryAsync();
 
                 //Edit Schueler sstufe [schueler]
-                connection.Open();
                 command =
                 new NpgsqlCommand
                 ($"update schueler set sstufe = {changedschueler.Schuljahr} where skennzahl = {id}; ", connection);
-                command.ExecuteNonQuery();
-
+                i += await command.ExecuteNonQueryAsync();
                 connection.Close();
+                if (i == 2)
+                    return true;
+                return false;
             }
         }
 
-        public static SchuelerDLModel GetByIdSchueler(int id)
+        public async static Task<SchuelerDLModel> GetByIdSchuelerAsync(string sql,int id)
         {
-            using (NpgsqlConnection connection = new NpgsqlConnection(SqlDataAccess.GetConnectionString()))
+            using (NpgsqlConnection connection = new NpgsqlConnection(sql))
             {
                 NpgsqlCommand command;
 
@@ -119,7 +120,7 @@ namespace Ubigrade.Library.Processors
                 $"join person on skennzahl = checkpersonnumber " +
                 $"where checkpersonnumber = {id} " +
                 $"order by skennzahl;", connection);
-                var dr = command.ExecuteReader();
+                var dr = await command.ExecuteReaderAsync();
                 SchuelerDLModel editschueler = null;
 
                 while (dr.Read())
@@ -132,34 +133,37 @@ namespace Ubigrade.Library.Processors
             }
         }
 
-        public static void DeleteSchueler(int id)
+        public static async Task<bool> DeleteSchuelerAsync(string sql,int id)
         {
-            using (NpgsqlConnection connection = new NpgsqlConnection(SqlDataAccess.GetConnectionString()))
+            using (NpgsqlConnection connection = new NpgsqlConnection(sql))
             {
                 NpgsqlCommand command;
-
+                //noch ned ganz fertig denk ich - rohit
                 //Delete Schueler props [person]
                 connection.Open();
                 command =
                     new NpgsqlCommand($"delete from person where checkpersonnumber = {id};", connection);
-                command.ExecuteNonQuery();
+                int i = await command.ExecuteNonQueryAsync();
+                int b = 0;
+                if (i == 1)
+                {
+                    command =
+                        new NpgsqlCommand($"delete from schueler where skennzahl = {id};", connection);
+                    b = await command.ExecuteNonQueryAsync();
+                    connection.Close();
 
-                connection.Close();
+                }
+                if (b == 1)
+                    return true;
+                return false;
 
-                //Delete sstufe [schueler]
-                connection.Open();
-                command =
-                    new NpgsqlCommand($"delete from schueler where skennzahl = {id};", connection);
-                command.ExecuteNonQuery();
-
-                connection.Close();
             }
         }
 
-        public static List<FaecherDLModel> LoadSchuelerFaecher(int id)
+        public async static Task<List<FaecherDLModel>> LoadSchuelerFaecherAsync(string sql,int id)
         {
             List<FaecherDLModel> SchuelerFaecher = new List<FaecherDLModel>();
-            using (NpgsqlConnection connection = new NpgsqlConnection(SqlDataAccess.GetConnectionString()))
+            using (NpgsqlConnection connection = new NpgsqlConnection(sql))
             {
                 SchuelerFaecher.Clear();
                 NpgsqlCommand command;
@@ -167,7 +171,7 @@ namespace Ubigrade.Library.Processors
                 connection.Open();
                 command =
                 new NpgsqlCommand($"select skennzahl, bez from schueler2faecher where skennzahl = {id} order by 2;", connection);
-                var dr = command.ExecuteReader();
+                var dr = await command.ExecuteReaderAsync();
 
                 while (dr.Read())
                 {
